+++ packages/state-manager/src/sagas/communities/communities.selectors.test.ts
@@ -44,49 +44,6 @@ describe('communitiesSelectors', () => {
     expect(community).toEqual({ ...communityAlpha })
   })
 
-  it('invitationUrl selector does not break if there is no community', () => {
-    const { store } = prepareStore()
-    const invitationUrl = communitiesSelectors.invitationUrl(store.getState())
-    expect(invitationUrl).toEqual('')
-  })
-
-  it('invitationUrl selector returns proper url', async () => {
-    const { store } = prepareStore()
-    const factory = await getFactory(store)
-    const peerList = [
-      createLibp2pAddress(
-        'gloao6h5plwjy4tdlze24zzgcxll6upq2ex2fmu2ohhyu4gtys4nrjad',
-        'QmZoiJNAvCffeEHBjk766nLuKVdkxkAT7wfFJDPPLsbKSA'
-      ),
-    ]
-    const psk = '12345'
-    const ownerOrbitDbIdentity = 'testOwnerOrbitDbIdentity'
-    await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>('Community', {
-      peerList,
-      ownerOrbitDbIdentity,
-    })
-    store.dispatch(communitiesActions.savePSK(psk))
-    const selectorInvitationUrl = communitiesSelectors.invitationUrl(store.getState())
-    const expectedUrl = invitationShareUrl(peerList, psk, ownerOrbitDbIdentity)
-    expect(expectedUrl).not.toEqual('')
-    expect(selectorInvitationUrl).toEqual(expectedUrl)
-  })
-
-  it('invitationUrl selector returns empty string if state lacks psk', async () => {
-    const { store } = prepareStore()
-    const factory = await getFactory(store)
-    await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>('Community', {
-      peerList: [
-        createLibp2pAddress(
-          'gloao6h5plwjy4tdlze24zzgcxll6upq2ex2fmu2ohhyu4gtys4nrjad',
-          'QmZoiJNAvCffeEHBjk766nLuKVdkxkAT7wfFJDPPLsbKSA'
-        ),
-      ],
-    })
-    const invitationUrl = communitiesSelectors.invitationUrl(store.getState())
-    expect(invitationUrl).toEqual('')
-  })
-
   it('returns proper ownerNickname - ownerCertificate exist', async () => {
     const { store } = prepareStore()
     const factory = await getFactory(store)