+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+[unreleased]
+
+# Fixes:
+
+* Make sure address of the inviting peer is in the invitation link
+
 [2.1.1]
 
 # Fixes: